+++ test/test_helper.rb
@@ -7,6 +7,7 @@ require_relative "../test/dummy/config/environment"
 ActiveRecord::Migrator.migrations_paths = [File.expand_path("../test/dummy/db/migrate", __dir__)]
 require "rails/test_help"
 require "rails/test_unit/reporter"
+require "pry"
 Rails::TestUnitReporter.executable = "bin/test"
 
 # Load fixtures from the engine