+++ CHANGELOG.md
@@ -1,9 +1,11 @@
-### 0.16
+### 0.17
+* Adding support for rails 7.2. Mostly this was fixing the deprecation of `request.show_exceptions?` and support the various
+  values of `Rails.application.config.action_dispatch.show_exceptions`.
 
+### 0.16
 * Ensure `decode_rack_response` always calls `#close` on the Rack body in tests, as some middleware might be applying a BodyProxy
 
 ### 0.15
-
 * Exclude versions of Rails 7 which were incompatible with the pbbuilder ActionView handler, as pbbuilder cannot work there at all
 * Fix decode_rack_response to be compatible with Rack response body wrappers (and conform to the Rack SPEC)